refactor(hf_client)!: move call_hf_pbr to utils layer
Move high-level call_hf_pbr function from hf_client.py to utils.py
to align with Platform client architecture pattern.

BREAKING CHANGE: call_hf_pbr is no longer exported from hf_client module.
Import from utils module instead.

- Move call_hf_pbr() from scripts/hf_client.py to scripts/utils.py
- Keep only low-level API helpers in hf_client.py (_resolve_fn_index,
  _upload_file, _join_queue, _poll_queue, _download_results, _download_file)
- Add required imports to utils.py (datetime, urllib, socket)
- Remove unused imports from hf_client.py (datetime, tempfile)

This separates concerns:
- hf_client.py: Low-level Gradio API client
- utils.py: High-level business logic wrappers

Related-To: #52
